requirements.txt
discord.py
py-cord
psutil
asyncio